refactor: consolidate overlapping tests
